turn fusionmanage into a ruby gem
